+++ CHANGELOG.md
@@ -11,6 +11,8 @@ please see [changelog_updates.md](docs/dev/changelog_updates.md).
 
 #### Minor
 
+- `Added wrapper API endpoint to query a single contract agreement
+
 #### Patch
 
 #### Major Changes